% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscedToYearsUS.R
\name{iscedToYearsUS}
\alias{iscedToYearsUS}
\title{Encodes ISCED Code as number of years of education for the United States.}
\usage{
iscedToYearsUS(responses)
}
\arguments{
\item{responses}{data.frame of student responses, including at least the
following columns:
  studentId
  iscedMother
  iscedFather}
}
\value{
encoded responses data.frame with columns:
  studentId
  yearsOfEducationMother
  yearsOfEducationFather
}
\description{
Derives years of education of student's parents
based on student's choice of ISCED codes. See:
http://uis.unesco.org/en/topic/international-standard-classification-education-isced
for more information about ISCED codes
}
